pyproject.toml
[tool.poetry]
name = "glpibot"
version = "0.1.0"
description = "Cloudate INFRA telegram bot glpi"
authors = ["nic <romb.intu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
glpi-api = "^0.5.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
uvicorn = {extras = ["classic"], version = "^0.32.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
